Create tic_tac_toe part 2
I've got the random start done, the writing of the 'X'/'O' done! The board is displayed with just X's & O's, and it displays the numbering of the slots from 1 - 9. I've not tested the 'if slot is already filled, and the game board display of each player1/2, 'X' or 'O'! There's also no 'winner' check function.